% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konto_check.R
\name{get_db_konto_list}
\alias{get_db_konto_list}
\title{Get current list of konto numbers in database}
\usage{
get_db_konto_list(con)
}
\arguments{
\item{con}{database connection}
}
\value{
table of kontos
}
\description{
Get current list of konto numbers in database
}
