% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konto_check.R
\name{get_konto_list_full}
\alias{get_konto_list_full}
\title{Get current list of valid konto numbers from newest EK file}
\usage{
get_konto_list_full(folder, file = NULL)
}
\arguments{
\item{folder}{folder path of the EK file}

\item{file}{file for testing}
}
\value{
table of kontos
}
\description{
Get current list of valid konto numbers from newest EK file
}
