% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_n_utilities.R
\name{get_most_recent_file_from_pattern}
\alias{get_most_recent_file_from_pattern}
\title{Get most recent file from regex pattern}
\usage{
get_most_recent_file_from_pattern(folder, pattern)
}
\arguments{
\item{folder}{folder to search}

\item{pattern}{regex pattern to match}
}
\value{
full name of file
}
\description{
Get most recent file from regex pattern
}
