% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{konto_codes}
\alias{konto_codes}
\title{Lookup for missing konto codes}
\usage{
konto_codes(x)
}
\arguments{
\item{x}{character input}
}
\value{
lookedup code
}
\description{
Switch function for missing konto codes.
}
\keyword{internal}
