% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{load_series_as_df}
\alias{load_series_as_df}
\title{Loading function}
\usage{
load_series_as_df(series_list, con)
}
\arguments{
\item{series_list}{list of character objects of series codes}

\item{con}{connection to database}
}
\value{
dataframe of all series in the list
}
\description{
Loads all the series from the list of series codes, by first getting the list
of the newest vintage numbers. Theg gets the data, renames the value column
into the series code
}
