% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{mf_excel_parser}
\alias{mf_excel_parser}
\title{MF Excel table parser}
\usage{
mf_excel_parser(file_path, table_name, sheet_name)
}
\arguments{
\item{file_path}{path to excel file}

\item{table_name}{code of table to be used in series codes}

\item{sheet_name}{the name of the sheet where the table is}
}
\value{
list of three tables: annual and monthly series and the series codelist
}
\description{
Parsing function to extract and clean up the budget timeseries data in
the Ministry of Finance excel spreadsheets usually published [here](https://www.gov.si/teme/fiskalna-in-javnofinancna-politika/).
The parser works on 5 tables with untidy budget account data, furthermore
the tables have errors and inconsistencies. The parser first extracts the header
because it has 2-3 rows, then the data. Cleaning up the empty rows, adding
missing codes, transposing and renaming the columns.
}
\details{
Hardcoded failure points:
* there is a single duplicated row, if more turn up, the funciton will fail
* the first row of the header is 8 rows above the first data row. If that
chages or the code of the first data row changes, this will fail.
* one of the tables has a six-month column instead of months and years. It
is removed, and presumably nothing similar can ever happen again.
}
