% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_n_utilities.R
\docType{data}
\name{mf_file_patterns}
\alias{mf_file_patterns}
\title{Regex patterns for MF export file names}
\format{
An object of class \code{character} of length 2.
}
\usage{
mf_file_patterns
}
\description{
Anchored to the timestamp so that ad-hoc exports such as
`Export_4BJF_ZZZZ_<timestamp>.csv` are not picked up.
}
\keyword{datasets}
