% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{month_codes}
\alias{month_codes}
\title{Lookup for month codes}
\usage{
month_codes(x)
}
\arguments{
\item{x}{character string with month name}
}
\value{
code
}
\description{
Switch function for month codes
}
\keyword{internal}
