% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prepare_annual_eo}
\alias{prepare_annual_eo}
\title{Calculate and prepare the table for annual data for EO}
\usage{
prepare_annual_eo(data_frame)
}
\arguments{
\item{data_frame}{output from  \link[MFfetchR]{transform_series_eo}}
}
\value{
list of three objects, the main table and two header rows.
}
\description{
This function aggregates the amounts over years (also for the last incomplete year)
then calculates the year-on-year growth rates, and the relative contributions
for each account and formats the table appropriately. It also prepares the
two header rows required for a pretty (but def not tidy) Excel presentation.
}
