% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_old.R
\name{prepare_category_relationship_table}
\alias{prepare_category_relationship_table}
\title{Prepare table to insert into `category_relationship` table}
\usage{
prepare_category_relationship_table(table_name, con, schema = "platform")
}
\arguments{
\item{table_name}{character string of table code}

\item{con}{connection to the database}

\item{schema}{schema name defaults to "platform"}
}
\value{
a dataframe with the `id`, `parent_id`, `source_id` for each relationship
betweeh categories
}
\description{
Helper function that manually prepares the category_relationship table.
Returns table ready to insert into the `category` table.
}
