% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_old.R
\name{prepare_category_table}
\alias{prepare_category_table}
\title{Prepare table to insert into `category` table}
\usage{
prepare_category_table(table_name, con, schema = "platform")
}
\arguments{
\item{table_name}{character string of table code}

\item{con}{connection to the database}

\item{schema}{schema name defaulting to "platform"}
}
\value{
a dataframe with the `id`, `name`, `source_id` for each category that
the table is a member of.
}
\description{
Helper function that extracts all the parent categories from the full
hierarchy data.frame, and prepares the category table with field ids and
their names. Actually duplicates the overall category, but who gives a shit.
}
