% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_old.R
\name{prepare_category_table_table}
\alias{prepare_category_table_table}
\title{Prepare table to insert into `category_table` table}
\usage{
prepare_category_table_table(table_name, con, schema = "platform")
}
\arguments{
\item{table_name}{character string of table code}

\item{con}{connection to the database}

\item{schema}{schema name defaults to "platform"}
}
\value{
a dataframe with the `category_id` `table_id` and `source_id` columns for
each table-category relationship.
}
\description{
Helper function that manually prepares the category_table table.
Returns table ready to insert into the `category` table.
}
