% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_new.R
\name{prepare_mf_data_for_insert_new}
\alias{prepare_mf_data_for_insert_new}
\title{#' Prepare vintage table for M or A
#'
#' Helper function preparing for the vintage table for a specific table
#' and either the monthly or the annual data. Uses current time as `published`
#'
#' @param interval "M" or "A"
#' @param tbl_id numeric table id
#' @param con connection to the database.
#'
#' @return data frame with `series_id` and `published` columns
#' @keywords internal}
\usage{
prepare_mf_data_for_insert_new(final_data, con, schema = "platform")
}
\arguments{
\item{con}{connection to database}

\item{schema}{schema name defaults to "parameter"}

\item{parsed_data}{list with at least monthly and annual dataframes with the data_points
output of \link[MFfetchR]{mf_excel_parser}.}
}
\value{
a dataframe with the period_id, value and id values for all the vintages in the table.
}
\description{
vintage_table <- function(interval, tbl_id, con) {
  dplyr::tbl(con, "series") %>%
    dplyr::filter(table_id == tbl_id,
                  interval_id == interval) %>%
    dplyr::select(series_id=id) %>%
    dplyr::collect() %>%
    dplyr::mutate(published = get_published_time())
}
}
\details{
Get and prepare data for import

Prepares the timeseries data for importing into the database. Only works after
vintages have been imported
}
