% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prepare_monthly_12mK}
\alias{prepare_monthly_12mK}
\title{Prepare data for the 12 month cumulative table}
\usage{
prepare_monthly_12mK(data_frame)
}
\arguments{
\item{data_frame}{output from  \link[MFfetchR]{transform_series_12mK}}
}
\value{
list of four tables: the first with the monthly data, the other three
are headers for the fancy excel three-row heaedr
}
\description{
This function calculates the 12 month cumulative sums and the preps the
table and headers for exporting into excel
}
