% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_new.R
\name{prepare_series_levels_table_new}
\alias{prepare_series_levels_table_new}
\title{Prepare table to insert into `series_levels` table}
\usage{
prepare_series_levels_table_new(table_name, con, schema = "platform")
}
\arguments{
\item{table_name}{the table name (eg "DP")}

\item{con}{connection to the database}

\item{schema}{schema name defaults to "platform"}
}
\value{
a dataframe with the `series_id`, `tab_dim_id`, `value` columns
all the series-level combinatins for this table.
}
\description{
Helper function that extracts the individual levels for each series and
gets the correct dimension id for each one and the correct series id to
keep with the constraints.
Returns table ready to insert into the `series_levels`table with the
db_writing family of functions.
}
