% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_series_table}
\alias{prepare_series_table}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table(file_path, table_name, sheet_name, con)
}
\arguments{
\item{table_name}{character string e.g. "DP"}

\item{con}{connection to the database}
}
\value{
a dataframe with the following columns: `name_long`, `code`,
`unit_id`, `table_id` and `interval_id`for each series in the table
well as the same number of rows as there are series
}
\description{
Prepare table to insert into `series` table
}
