% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_new.R
\name{prepare_series_table_new}
\alias{prepare_series_table_new}
\title{Prepare table to insert into `series` table}
\usage{
prepare_series_table_new(file_path, table_name, con, schema = "platform")
}
\arguments{
\item{file_path}{path to the excel file}

\item{table_name}{character string of table code}

\item{con}{connection to the database}

\item{schema}{schema name defaults to "platform"}
}
\value{
a dataframe with the following columns: `name_long`, `code`,
`unit_id`, `table_id` and `interval_id`for each series in the table
well as the same number of rows as there are series
}
\description{
Prepare table to insert into `series` table
}
