% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{prepare_stats_appendix}
\alias{prepare_stats_appendix}
\title{Prepare data for the stats appendix table}
\usage{
prepare_stats_appendix(data_frame)
}
\arguments{
\item{data_frame}{output from  \link[MFfetchR]{transform_series_12mK}}
}
\value{
list of four tables: the first with the monthly data, the other three
are headers for the fancy excel three-row heaedr
}
\description{
This function prepares the aggregates for the stats appendix in EO. This
means the last three complete years, the last 12 complete quaters and last
24 complete months.
}
