% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_old.R
\name{prepare_table_table}
\alias{prepare_table_table}
\title{Prepare table to insert into `table` table}
\usage{
prepare_table_table(table_name, keep_vintage = FALSE, con, schema = "platform")
}
\arguments{
\item{table_name}{character string of table code}

\item{keep_vintage}{boolean whether to keep vintages}

\item{con}{a connection to the database}

\item{schema}{the schema to use for the connection, default is "platform"}
}
\value{
a dataframe with the `code`, `name`, `source_id`, `url`, and `notes` columns
for this table.
}
\description{
Returns table ready to insert into the `table` table with the db_writing family
of functions.
}
