% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_structure.R
\name{prepare_unit_table}
\alias{prepare_unit_table}
\title{Prepare table to insert into `units` table}
\usage{
prepare_unit_table(con)
}
\arguments{
\item{con}{connection to the database}
}
\value{
a dataframe with the single column containing the different units used
in this table.
}
\description{
Helper function that manually prepares the unit row.
Returns table ready to insert into the `units`table with the
db_writing family of functions.
}
