% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{prepare_vintage_table}
\alias{prepare_vintage_table}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table(file_path, table_name, sheet_name, con)
}
\arguments{
\item{file_path}{path to excel file}

\item{table_name}{name of table}

\item{sheet_name}{name of sheet}

\item{con}{connection to database}
}
\value{
a dataframe with the `series_id` and `published` columns
for all the series in this table.
}
\description{
Helper function that prepares the vintage table with the new vintages.
First checks if new vintages are even necessary, by parsing the excel file
to get the most recent month, (and year?) and then checking with the most
recent period in the database. If there is new data,
Prepare a dataframe with the `series_id`s of the table and the current time
as their publication date - since we don't have anything better.
}
\details{
Returns table ready to insert into the `vintage`table with the
db_writing family of functions.
}
