% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_new.R
\name{prepare_vintage_table_and_merge_data_points}
\alias{prepare_vintage_table_and_merge_data_points}
\title{Prepare table to insert into `vintage` table}
\usage{
prepare_vintage_table_and_merge_data_points(
  file_path,
  file_name = NULL,
  table_name,
  con,
  schema = "platform"
)
}
\arguments{
\item{file_path}{path to excel file}

\item{file_name}{name of file for testing}

\item{table_name}{name of table}

\item{con}{connection to database}

\item{schema}{schema name defaults to "platform"}
}
\value{
list with a dataframe with the `series_id` and `published` columns
for all monthly and annual series in this table. and the parsed data.
}
\description{
Helper function that prepares the vintage table with the new vintages.
First checks if new vintages are even necessary, by parsing the excel file
to get the most recent month, (and year?) and then checking with the most
recent period in the database. If there is new data,
Prepare a dataframe with the `series_id`s of the table and the current time
as their publication date - since we don't have anything better.
}
\details{
Returns table ready to insert into the `vintage`table with the
db_writing family of functions.
}
