% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{rename_value_column}
\alias{rename_value_column}
\title{Helper to rename value column into seires name}
\usage{
rename_value_column(series_name, df)
}
\arguments{
\item{series_name}{character}

\item{df}{data frame with a column named `value`}
}
\value{
same dataframe, with renamed column
}
\description{
Helper to rename value column into seires name
}
\keyword{internal}
