% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_series_12mK}
\alias{transform_series_12mK}
\title{Transform series needed for the 12month cumulative tables}
\usage{
transform_series_12mK(raw_data_frame)
}
\arguments{
\item{raw_data_frame}{dataframe with series from the kbjf_series_list_12mK list}
}
\value{
dataframe with some additional series calculated
}
\description{
From the raw_data_frame output of load_series_as_df using kbjf_series_list_12mK as
the series list, this function calculates the necessary series and transforms
the values into millions.
}
