% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_inter}
\alias{trim_inter}
\title{Trim all spaces}
\usage{
trim_inter(x)
}
\arguments{
\item{x}{character string}
}
\value{
trimmed character string
}
\description{
Removes all spaces
}
\keyword{internal}
