% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_leading}
\alias{trim_leading}
\title{Trim leading spaces}
\usage{
trim_leading(x)
}
\arguments{
\item{x}{character string}
}
\value{
trimmed character string
}
\description{
Trim leading spaces
}
\keyword{internal}
