% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_setup_n_utilities.R
\name{update_JF_lookup_table_on_db}
\alias{update_JF_lookup_table_on_db}
\title{Update konto lookup table on database when new kontos arrive}
\usage{
update_JF_lookup_table_on_db(file_path)
}
\arguments{
\item{file_path}{what it says on the tin}
}
\value{
nothin
}
\description{
This funciton is called in the main script under the condition that
new kontos have been added. it recreates the konto lookup table on the
database, which is used in the joins for the materialised views
}
