% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{vintage_table}
\alias{vintage_table}
\title{Prepare vintage table for M or A}
\usage{
vintage_table(interval, tbl_id, con)
}
\arguments{
\item{interval}{"M" or "A"}

\item{tbl_id}{numeric table id}

\item{con}{connection to the database.}
}
\value{
data frame with series_id and published columns
}
\description{
Helper funciton preparing for the vintage table for a specific table
and either the montlhy or the annual data. Uses current time as `published`
}
\keyword{internal}
