% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{write_excel_kbjf_12mK}
\alias{write_excel_kbjf_12mK}
\title{Write 12mK to Excel}
\usage{
write_excel_kbjf_12mK(monthly_list, data_frame, outfile, update = TRUE)
}
\arguments{
\item{monthly_list}{list output of \link[MFfetchR]{prepare_monthly_12mK}}

\item{data_frame}{list output of \link[MFfetchR]{transform_series_eo}}

\item{outfile}{path of file to create}

\item{update}{logical whether file exists and should be loaded first and then
updated.}
}
\value{
nothing, creates excel file
}
\description{
Writes and formats the in a suitable format. Potentially lets you extend
a plot.
}
