% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{write_excel_kbjf_eo}
\alias{write_excel_kbjf_eo}
\title{Write EO tables to Excel}
\usage{
write_excel_kbjf_eo(quarterly_list, annual_list, data_frame, outfile)
}
\arguments{
\item{quarterly_list}{list output of \link[MFfetchR]{prepare_quarterly_eo}}

\item{annual_list}{list output of \link[MFfetchR]{prepare_annual_eo}}

\item{data_frame}{list output of \link[MFfetchR]{transform_series_eo}}

\item{outfile}{path of file to create}
}
\value{
nothing, creates excel file
}
\description{
Writes and formats the quarterly and annual and original raw data in a way
suitable for the EO publication.
}
