% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling.R
\name{write_excel_stats_appendix}
\alias{write_excel_stats_appendix}
\title{Write stats appendix to Excel}
\usage{
write_excel_stats_appendix(stats_appendix_list, data_frame, outfile)
}
\arguments{
\item{stats_appendix_list}{list output of \link[MFfetchR]{prepare_stats_appendix}}

\item{data_frame}{list output of \link[MFfetchR]{transform_series_12mK}}

\item{outfile}{path of file to create}
}
\value{
nothing, creates excel file
}
\description{
Writes and formats the in a suitable format.
}
