% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konto_check.R
\name{check_for_extra_kontos}
\alias{check_for_extra_kontos}
\title{Check for new kontos}
\usage{
check_for_extra_kontos(folder, con)
}
\arguments{
\item{folder}{folder path of the 4BJF file}

\item{con}{database connection}
}
\value{
logical for new kotnos
}
\description{
Check for new kontos
}
