% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/konto_check.R
\name{get_konto_list_data}
\alias{get_konto_list_data}
\title{Get current list of delivered konto numbers from newest 4BJF file}
\usage{
get_konto_list_data(folder)
}
\arguments{
\item{folder}{folder path of the 4BJF file}
}
\value{
table of kontos
}
\description{
Get current list of delivered konto numbers from newest 4BJF file
}
