% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_new.R
\name{insert_data_points_new}
\alias{insert_data_points_new}
\title{Insert datapoints into data_point table}
\usage{
insert_data_points_new(final_data, con, schema = "platform")
}
\arguments{
\item{con}{connection to database}

\item{schema}{schema name defaults to "platform"}

\item{parsed_data}{list with at least monthly and annual dataframes with the data_points
output of \link[MFfetchR]{mf_excel_parser}.}
}
\value{
nothing, just some printing along the way
}
\description{
So, the function extracts and preps the data with \link[MFfetchR]{prepare_mf_data_for_insert}
and writes it to a temporary table in the database.
}
\details{
It inserts any new periods into the period table,
adds the data points to the data point table.
}
