% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excelling_new.R
\name{mf_csv_parser_new}
\alias{mf_csv_parser_new}
\title{MF csv table parser}
\usage{
mf_csv_parser_new(file_path, file_name = NULL)
}
\arguments{
\item{file_path}{path to csv file}

\item{file_name}{name of file for testing}
}
\value{
list of three tables: annual and monthly series and the series codelist
}
\description{
This function reads the data from the CSV file and processes it.
It filters out rows with MESEC == 0, modifies BLG_ID, and creates new columns K1 to K4.
It also creates a period_id column and checks for duplications in the data.
Finally, it aggregates the data at different levels (K1, K2, K3, K4) and returns a final dataframe.
In addition it calculates all the 9XX transformations.
}
