% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_new.R
\name{MF_import_structure_new}
\alias{MF_import_structure_new}
\title{Import structural metadata for a MF table new}
\usage{
MF_import_structure_new(
  file_path,
  encoding = "UTF-16LE",
  table_name,
  con,
  schema = "platform",
  keep_vintage = FALSE
)
}
\arguments{
\item{file_path}{file path to data file}

\item{encoding}{what it says on the tin}

\item{table_name}{the table name (eg "DP")}

\item{con}{connection to database}

\item{schema}{schema name, defaults to "platform"}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}
}
\value{
nothing
}
\description{
Umbrella function that prepares and import all the metadata tables into
the database. It uses the functions from the UMARimportR package to
insert the data into the database. But is used only for adding new
series to the existing database from the new formatted csv files.
}
