% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importing_old.R
\name{MF_import_structure_old}
\alias{MF_import_structure_old}
\title{Import structural metadata for a BS table}
\usage{
MF_import_structure_old(
  table_name,
  con,
  schema = "platform",
  keep_vintage = FALSE,
  data_path
)
}
\arguments{
\item{table_name}{the table name (eg "DP")}

\item{con}{connection to database}

\item{schema}{schema name, defaults to "platform"}

\item{keep_vintage}{logical indicating whether to keep vintages, defaults to F}

\item{data_path}{path to where the data is.}
}
\value{
nothing
}
\description{
Umbrella function that prepares and import all the metadata tables into
the database. It uses the functions from the UMARimportR package to
insert the data into the database.
}
