% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_old.R
\name{prepare_dimension_levels_table}
\alias{prepare_dimension_levels_table}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table(
  file_path,
  table_name,
  sheet_name,
  con,
  schema = "platform"
)
}
\arguments{
\item{file_path}{path to the excel file}

\item{table_name}{character string of table code}

\item{sheet_name}{character string of sheet name with table in excel.}

\item{con}{connection to the database}

\item{schema}{schema name defaults to "platform"}
}
\value{
a dataframe with the `dimension_id`, `values` and `valueTexts`
columns for this table.
}
\description{
Helper function that manually prepares the dimension_levels for each
table and get their codes and text.
Returns table ready to insert into the `dimension_levels`table with the
db_writing family of functions.
}
