% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_structure_new.R
\name{prepare_dimension_levels_table_new}
\alias{prepare_dimension_levels_table_new}
\title{Prepare table to insert into `dimension_levels` table}
\usage{
prepare_dimension_levels_table_new(
  file_path,
  encoding = "UTF-16LE",
  table_name,
  con,
  schema = "platform"
)
}
\arguments{
\item{file_path}{path to the csv file}

\item{encoding}{of the file}

\item{table_name}{character string of table code}

\item{con}{connection to the database}

\item{schema}{schema name defaults to "platform"}
}
\value{
a dataframe with the `tab_dim_id`, `level_value` and `level_text`
columns for this table.
}
\description{
Helper function that manually prepares the dimension_levels for each
table and get their codes and text.
Returns table ready to insert into the `dimension_levels`table with the
db_writing family of functions.
}
