% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_new.R
\name{prepare_mf_data_for_insert_new}
\alias{prepare_mf_data_for_insert_new}
\title{Get and prepare data for import}
\usage{
prepare_mf_data_for_insert_new(final_data, con, schema = "platform")
}
\arguments{
\item{final_data}{list with at least monthly and annual dataframes with the data_points
output of \link[MFfetchR]{prepare_vintage_table_and_merge_data_points}.}

\item{con}{connection to database}

\item{schema}{schema name defaults to "parameter"}
}
\value{
a dataframe with the period_id, value and id values for all the vintages in the table.
}
\description{
Prepares the timeseries data for importing into the database. Only works after
vintages have been imported
}
